requirements.txt
aiogram>=3.0.0b8
asyncpg>=0.30.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
apscheduler>=3.10.1